feat: Add database setup script
The code changes include adding a PHP script to set up the database. The script creates a connection to the MySQL server, creates a new database, selects the database, and creates a table. This script will be useful for setting up the database for the application.

Please note that the commit message should not include any meta information like issue references, tags, or author names.